[WIP] block body scroll on mobile search
